Reorganizes conftest code
