+++ integration/tests/__init__.py
@@ -1,1 +1,4 @@
 # -*- coding: utf-8 -*-
+import logging
+
+logging.basicConfig(format='%(asctime)s [%(levelname)s] [%(process)d] %(message)s', level=logging.DEBUG)
