--- integration/tests/cook/__init__.py
@@ -1,3 +1,0 @@
-import logging
-
-logging.basicConfig(format='%(asctime)s [%(levelname)s] %(message)s', level=logging.DEBUG)
